# frontend/openapi.json
{
  "components": {
    "schemas": {
      "ActionClass": {
        "description": "Candidate action classes defined by the intervention knowledge base.\n\nThe LLM does not extend this list.\nNew classes are added to the knowledge base with a cited source.",
        "enum": [
          "monitoring",
          "targeted_disease_survey",
          "biosecurity_workflow",
          "water_quality_investigation",
          "physical_damage_assessment",
          "temporary_shading"
        ],
        "title": "ActionClass",
        "type": "string"
      },
      "Assignment": {
        "description": "One boat and team sent to one site to perform one action.",
        "properties": {
          "action_class": {
            "$ref": "#/components/schemas/ActionClass"
          },
          "action_id": {
            "title": "Action Id",
            "type": "string"
          },
          "boat_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boat Id"
          },
          "compatibility_rationale": {
            "description": "Why this action was considered compatible, with its knowledge-base provenance.",
            "title": "Compatibility Rationale",
            "type": "string"
          },
          "estimated_cost_usd": {
            "minimum": 0,
            "title": "Estimated Cost Usd",
            "type": "number"
          },
          "estimated_hours": {
            "minimum": 0,
            "title": "Estimated Hours",
            "type": "number"
          },
          "evidence_summary": {
            "description": "What evidence supported this decision.",
            "title": "Evidence Summary",
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "remaining_uncertainty": {
            "description": "What we still do not know.",
            "title": "Remaining Uncertainty",
            "type": "string"
          },
          "requires_manager_approval": {
            "default": true,
            "title": "Requires Manager Approval",
            "type": "boolean"
          },
          "site_id": {
            "title": "Site Id",
            "type": "string"
          },
          "site_name": {
            "title": "Site Name",
            "type": "string"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          }
        },
        "required": [
          "site_id",
          "site_name",
          "action_id",
          "action_class",
          "priority",
          "estimated_hours",
          "estimated_cost_usd",
          "evidence_summary",
          "remaining_uncertainty",
          "compatibility_rationale"
        ],
        "title": "Assignment",
        "type": "object"
      },
...
      "DeferredSite": {
        "description": "A site that needed attention but did not fit within capacity.",
        "properties": {
          "fallback_action_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "For example monitoring only, when intervention was deferred.",
            "title": "Fallback Action Id"
          },
          "reason": {
            "description": "The binding constraint, stated plainly.",
            "title": "Reason",
            "type": "string"
          },
          "site_id": {
            "title": "Site Id",
            "type": "string"
          },
          "site_name": {
            "title": "Site Name",
            "type": "string"
          }
        },
        "required": [
          "site_id",
          "site_name",
          "reason"
        ],
        "title": "DeferredSite",
        "type": "object"
      },
...
      "FieldReport": {
        "description": "A raw observation as submitted by a diver, scientist, or restoration team.",
        "properties": {
          "image_refs": {
            "items": {
              "type": "string"
            },
            "title": "Image Refs",
            "type": "array"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "observer": {
            "title": "Observer",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance",
            "default": "synthetic",
            "description": "Demo reports are synthetic and labeled as such everywhere they appear."
          },
          "provenance_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProvenanceMetadata"
              },
              {
                "type": "null"
              }
            ],
            "description": "Record-level source and retrieval metadata from the fixture envelope."
          },
          "report_id": {
            "title": "Report Id",
            "type": "string"
          },
          "site_id": {
            "title": "Site Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "report_id",
          "site_id",
          "observed_at",
          "observer",
          "text"
        ],
        "title": "FieldReport",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ObservationAccepted": {
        "additionalProperties": false,
        "description": "Result of submitting one field report, including the plan it produced.",
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/ResponsePlan"
          },
          "replan_latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Measured server-side so the dashboard does not include network noise.",
            "title": "Replan Latency Ms"
          },
          "report_id": {
            "title": "Report Id",
            "type": "string"
          }
        },
        "required": [
          "report_id",
          "plan"
        ],
        "title": "ObservationAccepted",
        "type": "object"
      },
      "Priority": {
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "Priority",
        "type": "string"
      },
      "Provenance": {
        "description": "Where a value actually came from.\n\nEvery externally sourced value carries one of these.\nSIMULATED and SYNTHETIC are never presented to a user as real.",
        "enum": [
          "live",
          "cache",
          "simulated",
          "synthetic"
        ],
        "title": "Provenance",
        "type": "string"
      },
      "ProvenanceMetadata": {
        "additionalProperties": false,
        "description": "Structured origin metadata for one value or coherent record.\n\n`fetched_at` records when an external value was retrieved.\n`observed_at` records when the source says the condition occurred.\nThey are deliberately separate because snapshot age and observation age are\ndifferent facts.",
        "properties": {
          "fetched_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At"
          },
          "kind": {
            "$ref": "#/components/schemas/Provenance"
          },
          "note": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "observed_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "source": {
            "description": "Human-readable source or fixture author.",
            "minLength": 1,
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          }
        },
        "required": [
          "kind",
          "source"
        ],
        "title": "ProvenanceMetadata",
        "type": "object"
      },
...
      "ReportStructureResult": {
        "additionalProperties": false,
        "description": "Validated observation extracted from one raw report by the configured LLM.",
        "properties": {
          "attempt_count": {
            "minimum": 1,
            "title": "Attempt Count",
            "type": "integer"
          },
          "extraction_confidence": {
            "maximum": 1,
            "minimum": 0,
            "title": "Extraction Confidence",
            "type": "number"
          },
          "input_tokens": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "latency_ms": {
            "minimum": 0,
            "title": "Latency Ms",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "observation": {
            "$ref": "#/components/schemas/StructuredObservation"
          },
          "output_tokens": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "report_id": {
            "title": "Report Id",
            "type": "string"
          }
        },
        "required": [
          "report_id",
          "observation",
          "extraction_confidence",
          "provider",
          "model",
          "attempt_count",
          "latency_ms"
        ],
        "title": "ReportStructureResult",
        "type": "object"
      },
...
      "ResponsePlan": {
        "description": "A complete allocation for one planning window.",
        "properties": {
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/Assignment"
            },
            "title": "Assignments",
            "type": "array"
          },
          "binding_constraints": {
            "description": "Which constraints were tight, so trade-offs are explainable.",
            "items": {
              "type": "string"
            },
            "title": "Binding Constraints",
            "type": "array"
          },
          "deferred": {
            "items": {
              "$ref": "#/components/schemas/DeferredSite"
            },
            "title": "Deferred",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "plan_id": {
            "title": "Plan Id",
            "type": "string"
          },
          "replan_latency_ms": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Evidence submitted to plan displayed. Reported metric, see docs/evaluation.md.",
            "title": "Replan Latency Ms"
          },
          "replan_trigger": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What caused this recompute, when it was not the first plan.",
            "title": "Replan Trigger"
          },
          "scenario_banner": {
            "description": "Simulated-data banner text, carried on the plan so it cannot be dropped.",
            "title": "Scenario Banner",
            "type": "string"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "total_strategic_value": {
            "description": "Objective value achieved. Comparable against the naive baseline.",
            "minimum": 0,
            "title": "Total Strategic Value",
            "type": "number"
          }
        },
        "required": [
          "plan_id",
          "generated_at",
          "scenario_id",
          "scenario_banner",
          "assignments",
          "total_strategic_value"
        ],
        "title": "ResponsePlan",
        "type": "object"
      },
...
      "StructuredObservation": {
        "description": "Structured signals extracted from one field report.\n\nEvery field is optional because a real report rarely mentions everything.\nAbsence means \"not reported\", never \"reported as zero\".",
        "properties": {
          "affected_taxa": {
            "items": {
              "type": "string"
            },
            "title": "Affected Taxa",
            "type": "array"
          },
          "bleaching_pct": {
            "anyOf": [
              {
                "maximum": 100,
                "minimum": 0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bleaching Pct"
          },
          "broken_coral_observed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broken Coral Observed"
          },
          "compared_to_previous_dive": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compared To Previous Dive"
          },
          "extraction_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What the extractor was unsure about. Surfaced, not swallowed.",
            "title": "Extraction Notes"
          },
          "lesion_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lesion Description"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "paling_pct": {
            "anyOf": [
              {
                "maximum": 100,
                "minimum": 0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paling Pct"
          },
          "provenance_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProvenanceMetadata"
              },
              {
                "type": "null"
              }
            ],
            "description": "Record-level source and retrieval metadata from the fixture envelope."
          },
          "report_id": {
            "title": "Report Id",
            "type": "string"
          },
          "sediment_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sediment Note"
          },
          "site_id": {
            "title": "Site Id",
            "type": "string"
          },
          "spatial_progression": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "For example 'spreading westward since last dive'.",
            "title": "Spatial Progression"
          },
          "tissue_loss_observed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tissue Loss Observed"
          },
          "turbidity_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turbidity Note"
          }
        },
        "required": [
          "report_id",
          "site_id",
          "observed_at"
        ],
        "title": "StructuredObservation",
        "type": "object"
      },
      "StructuredObservationSubmission": {
        "additionalProperties": false,
        "description": "A raw report and the reviewed extraction to use for re-planning.",
        "properties": {
          "observation": {
            "$ref": "#/components/schemas/StructuredObservation"
          },
          "report": {
            "$ref": "#/components/schemas/FieldReport"
          }
        },
        "required": [
          "report",
          "observation"
        ],
        "title": "StructuredObservationSubmission",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Decision support that turns environmental monitoring, field observations, scientific intervention guidance, and limited conservation resources into continuously updated reef-response plans.",
    "title": "ReefCommand",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/observations": {
      "post": {
        "description": "Accept a field report, structure it, and trigger re-planning.",
        "operationId": "submit_observation_observations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Observation",
        "tags": [
          "observations"
        ]
      }
    },
    "/observations/structure": {
      "post": {
        "description": "Use the configured LLM to convert messy report text into a validated observation.",
        "operationId": "structure_report_observations_structure_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportStructureResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Structure Report",
        "tags": [
          "observations"
        ]
      }
    },
    "/observations/structured": {
      "post": {
        "description": "Accept a reviewed structured observation and trigger re-planning without another LLM call.",
        "operationId": "submit_structured_observation_observations_structured_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructuredObservationSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Structured Observation",
        "tags": [
          "observations"
        ]
      }
    },
...
  }
}